refactor: rename DataPoint to Event
BREAKING CHANGE: Rename `DataPoint` to `Event` so it is more in line
with the overall terminology used with the databuffer's backend.
If you specifically used the type name in your code, then you will need
to change it accordingly. If you only used type inference, you don't
need to change your code.